# nacl-collection/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /nacl/{id}:
    # our controller name
    x-swagger-router-controller: nacl
    get:
      operationId: getOne
      description: get a nacl
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetNaclResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: update
      description: update a nacl
      # define the parameters
      parameters:
        - name: id
          description: Nacl id
          type: string
          in: path
          required: true
        - name: ruleNumber
          description: Nacl properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Nacl"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: delNacl
      description: delete a nacl
      # define the parameters
      parameters:
        - name: id
          description: Nacl id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  GetNaclResponse:
      required:
        - id
        - cidrBlock
        - egress
        - fromPort
        - toPort
        - protocol
        - ruleAction
        - ruleNumber
      properties:
        id:
          type: string
        cidrBlock: 
          type: string
        egress:
          type: number
        fromPort:
          type: number
        toPort:
          type: number
        protocol:
          type: number
        ruleAction:
          type: string
        ruleNumber:
          type: number
  Nacl:
    type: object
    properties:
      cidrBlock: 
        type: string
        description: task object name
      egress:
        type: number
        description: task object name
      fromPort:
        type: number
        description: task object name
      toPort:
        type: number
        description: task object name
      protocol:
        type: number
        description: task object name
      ruleAction:
        type: string
        description: task object name
      ruleNumber:
        type: number
        description: task object name
    required:
      - cidrBlock
      - egress
      - fromPort
      - toPort
      - protocol
      - ruleAction
      - ruleNumber
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description